feat: add completer option to forward to readline (#51)
Author: @131

This is a rebase of https://github.com/npm/read/pull/28 with the test
removed.  Testing this requires installing an entirely new pseudo-tty
library, and given that the other flags aren't tested that seems like a
lot to ask at this point.  We need to readdress testing in general in
this repo but this PR is not the place to do it.

Co-authored-by: François Leurent <131.php@cloudyks.org>